fix: hoist toast host into bottom sheet
