+++ changelog.d/next/1165.fixed.md
@@ -1,1 +1,1 @@
-The wrong-passphrase error toast is now visible on top of the hardware transfer passphrase sheet.
+Error toasts are now visible on top of modal bottom sheets, including the hardware transfer passphrase sheet.
